Add single-threaded feature to librasan
