+++ libafl_qemu/librasan/asan/src/symbols/atomic_guest_addr/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "single-threaded")]
+pub mod single_threaded;
+
+#[cfg(not(feature = "single-threaded"))]
+pub mod multi_threaded;
