+++ libafl_qemu/librasan/qasan/Cargo.toml
@@ -8,7 +8,9 @@ rust-version.workspace = true
 crate-type = ["staticlib"]
 
 [features]
-default = []
+default = ["single-threaded"]
+## Support single-threaded targets by using dummy synchronization primitives
+single-threaded = ["dep:nospin", "asan/single-threaded"]
 test = ["asan/test", "dummy_libc/test"]
 
 [dependencies]
@@ -25,6 +27,11 @@ dummy_libc = { path = "../dummy_libc", default-features = false }
 libc = { version = "0.2.169", default-features = false }
 log = { version = "0.4.22", default-features = false, features = [
   "release_max_level_info",
+] }
+nospin = { version = "0.2.4", default-features = false, optional = true, features = [
+  "lazy",
+  "mutex",
+  "once",
 ] }
 spin = { version = "0.9.8", default-features = false, features = [
   "lazy",
