+++ libafl_qemu/librasan/zasan/src/lib.rs
@@ -19,6 +19,9 @@ use asan::{
     tracking::{Tracking, guest_fast::GuestFastTracking},
 };
 use log::{Level, trace};
+#[cfg(feature = "single-threaded")]
+use nospin::{Lazy, Mutex};
+#[cfg(not(feature = "single-threaded"))]
 use spin::{Lazy, Mutex};
 
 pub type ZasanFrontend = DefaultFrontend<
